// frontend/src/api/models/BonusView.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BonusType } from './BonusType'
export type BonusView = {
  id: string
  employee_id: string
  bonus_type: BonusType
  amount: string
  currency: string
  payment_date: string
  description: string | null
  created_at: string | null
  updated_at: string | null
}