// frontend/src/api/models/EmployeeListView.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EmploymentStatusType } from './EmploymentStatusType'
export type EmployeeListView = {
  id: string
  first_name: string
  last_name: string
  email: string
  hire_date: string | null
  current_status: EmploymentStatusType | null
}